+++ bazel/BUILD
@@ -36,9 +36,9 @@ cc_library(
         "ngraph_bridge/ngraph_enter_prefetch_in_catalog.h",
         "ngraph_bridge/ngraph_executor.h",
         "ngraph_bridge/ngraph_encapsulate_op.h",
+        "ngraph_bridge/ngraph_encapsulate_op_utils.h",
         "ngraph_bridge/ngraph_data_cache.h",
         "ngraph_bridge/ngraph_find_replace_prefetchdataset.h",
-        "ngraph_bridge/ngraph_encapsulate_get_prefetch.h",
         "ngraph_bridge/ngraph_freshness_tracker.h",
         "ngraph_bridge/ngraph_mark_for_clustering.h",
         "ngraph_bridge/ngraph_partial_shapes.h",
@@ -80,9 +80,9 @@ cc_library(
         "ngraph_bridge/ngraph_encapsulate_clusters.cc",
         "ngraph_bridge/ngraph_encapsulate_impl.cc",
         "ngraph_bridge/ngraph_enter_prefetch_in_catalog.cc",
-        "ngraph_bridge/ngraph_encapsulate_get_prefetch.cc",
         "ngraph_bridge/ngraph_executor.cc",
         "ngraph_bridge/ngraph_encapsulate_op.cc",
+        "ngraph_bridge/ngraph_encapsulate_op_utils.cc",
         "ngraph_bridge/ngraph_freshness_tracker.cc",
         "ngraph_bridge/ngraph_mark_for_clustering.cc",
         "ngraph_bridge/ngraph_partial_shapes.cc",
