+++ ngraph_bridge/CMakeLists.txt
@@ -48,6 +48,7 @@ set(SRC
    ngraph_executor.cc
    ops/ngraph_ops.cc
    ngraph_encapsulate_op.cc
+   ngraph_encapsulate_op_utils.cc
    ngraph_freshness_tracker.cc
    ngraph_mark_for_clustering.cc
    ngraph_partial_shapes.cc
@@ -60,7 +61,6 @@ set(SRC
    tf_deadness_analysis.cc
    prefetch_autotuner.cc
    ngraph_prefetch_dataset_op.cc
-   ngraph_encapsulate_get_prefetch.cc
    stats_utils.cc
    version.cc
 )
